Move tabs from Configuartion to page to their own components
